requirements.txt
streamlit
Pillow 